CRUD(insercion) en las tablas Ingreso e IngresoDetalle y CRUD completo de la tabla Cliente

+++ SWADNETVENTAS/App_Code/Entidades/EVDetalleIngreso.cs
@@ -20,7 +20,7 @@ public class EVDetalleIngreso
     [DataMember]
     public int CantidadIngreso { get; set; }
     [DataMember]
-    public decimal PrecioIngreso { get; set; }
+    public double PrecioIngreso { get; set; }
     #endregion
 
     #region Constructores
@@ -30,7 +30,7 @@ public class EVDetalleIngreso
         CodigoIngreso = int.MinValue;
         CodigoArticulo = int.MinValue;
         CantidadIngreso = int.MinValue;
-        PrecioIngreso = decimal.MinValue;
+        PrecioIngreso = double.MinValue;
     }
     #endregion
 }
